+++ src/backend/users/users.module.ts
@@ -5,10 +5,11 @@ import { AuthModule } from "../auth/auth.module";
 import { PasswordResetModule } from "../password-reset/password-reset.module";
 import { DbModule } from "../db/db.module";
 import { InstanceConfigModule } from "../instance-config/instance-config.module";
+import { UIStateService } from "./ui-state.service";
 
 @Module({
   imports: [DbModule, AuthModule, PasswordResetModule, InstanceConfigModule],
-  providers: [UsersService],
+  providers: [UsersService, UIStateService],
   controllers: [UsersController],
   exports: [UsersService],
 })