+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v0.39.0] - 2020-05-27
+
 - Fix bug where users were logged out randomly, b/c JWT secrets were rotated too often
 - Add support for storing class filters, synchronized across devices ([#380](https://gitlab.com/Skn0tt/EntE/-/issues/380))
 - Added links to docs to some administration pages ([#381](https://gitlab.com/Skn0tt/EntE/-/issues/381))